requirements.txt
pandas==0.25.1
readtime==1.1.1
en_core_web_sm==2.0.0
spacy==2.3.0
fuzzywuzzy==0.18.0
wordcloud==1.7.0
textblob==0.15.3
httpx==0.13.3
streamlit==0.76.0
pandas==1.1.5
spacy==3.0.1
nltk==3.5
sumy==0.8.1
wordcloud==1.8.1
textblob==0.15.3
gensim==3.8.3
readtime==1.1.1
en_core_web_sm==3.0.0
matplotlib==3.3.4
httpcore==0.9.1
docx==0.2.4
fuzzywuzzy==0.18.0
Pillow==8.1.0
pyLDAvis==3.1.0
pytest==6.2.2
